Split greynoise package into different modules
