+++ src/greynoise/exceptions.py
@@ -1,0 +1,6 @@
+class RequestFailure(Exception):
+    """Exception to capture a failed request."""
+
+
+class InvalidResponse(Exception):
+    """Exception to capture a failed response parse."""
